+++ examples/test_data_gen/src/test_data_gen_local/run_test_data_gen.py
@@ -1,43 +1,39 @@
-import json
 import os
 from datetime import datetime
 
 import configargparse
-from constants import TEXT_CHUNK
-from doc_split import split_doc
 
 from promptflow import PFClient
 from promptflow.entities import Run
+
+UTILS_PATH = os.path.abspath(os.path.join(os.getcwd(), "src", "utils"))
+if UTILS_PATH not in os.sys.path:
+    os.sys.path.insert(0, UTILS_PATH)
+
+from constants import TEXT_CHUNK, CONNECTIONS_TEMPLATE
+from common import split_document, clean_data_and_save
 
 CONFIG_FILE = os.path.abspath(os.path.join(os.path.dirname(__file__), "config.ini"))
 
 
 def batch_run_flow(
-    pf: PFClient,
-    flow_folder: str,
-    flow_input_data: str,
-    flow_batch_run_size: int,
-    connection_name: str = "azure_open_ai_connection",
+        pf: PFClient,
+        flow_folder: str,
+        flow_input_data: str,
+        flow_batch_run_size: int,
+        connection_name: str = "azure_open_ai_connection",
 ):
-    environment_variables = {
-        "PF_WORKER_COUNT": str(flow_batch_run_size),
-        "PF_BATCH_METHOD": "spawn",
-    }  # TODO: what does 'spawn' mean?
-
     print("start to run batch flow run.")
-    # create run
     base_run = pf.run(
         flow=flow_folder,
         data=flow_input_data,
-        stream=True,  # TODO: understand 'stream'
-        environment_variables=environment_variables,
-        connections={
-            "validate_and_generate_seed_question": {"connection": connection_name},
-            "validate_and_generate_test_question": {"connection": connection_name},
-            "validate_test_question": {"connection": connection_name},
-            "generate_ground_truth": {"connection": connection_name},
-            "validate_ground_truth": {"connection": connection_name},
+        stream=True,
+        environment_variables={
+            "PF_WORKER_COUNT": str(flow_batch_run_size),
+            "PF_BATCH_METHOD": "spawn",
         },
+        connections={key: {"connection": value["connection"].format(connection_name=connection_name)}
+                     for key, value in CONNECTIONS_TEMPLATE.items()},
         column_mapping={TEXT_CHUNK: "${data.text_chunk}"},
         debug=True,
     )
@@ -56,20 +52,6 @@ def get_batch_run_output(pf: PFClient, base_run: Run):
     ground_truth = details["outputs.ground_truth"].tolist()
     debug_info = details["outputs.debug_info"].tolist()
     return [{"question": q, "ground_truth": g, "debug_info": d} for q, g, d in zip(question, ground_truth, debug_info)]
-
-
-def clean_data_and_save(test_data_set: list, test_data_output_path: str):
-    cleaned_data = [
-        test_data
-        for test_data in test_data_set
-        if (test_data and all(val for key, val in test_data.items() if key.lower() != "line_number"))
-    ]
-
-    jsonl_str = "\n".join(map(json.dumps, cleaned_data))
-
-    cur_time_str = datetime.now().strftime("%b-%d-%Y-%H-%M-%S")
-    with open(os.path.join(test_data_output_path, "test-data-" + cur_time_str + ".jsonl"), "wt") as text_file:
-        print(f"{jsonl_str}", file=text_file)
 
 
 if __name__ == "__main__":
@@ -101,17 +83,26 @@ if __name__ == "__main__":
 
     # check_file_path_exists(args.test_data_output_path)
     if not args.should_skip_doc_split:
-        split_doc(args.documents_folder, args.document_nodes_output_path, args.document_chunk_size)
+        if not os.path.exists(args.document_nodes_output_path):
+            os.makedirs(args.document_nodes_output_path)
+
+        output = split_document(args.document_chunk_size, args.documents_folder, args.document_nodes_output_path)
 
     pf = PFClient()
     # TODO: error handling
     batch_run = batch_run_flow(
         pf,
         args.flow_folder,
-        args.document_nodes_output_path,
+        output,
         args.flow_batch_run_size,
         connection_name=args.connection_name,
     )
 
     test_data_set = get_batch_run_output(pf, batch_run)
-    clean_data_and_save(test_data_set, args.test_data_output_path)
+
+    cur_time_str = datetime.now().strftime("%b-%d-%Y-%H-%M-%S")
+    if not os.path.exists(args.test_data_output_path):
+        os.makedirs(args.test_data_output_path)
+
+    output = os.path.join(args.test_data_output_path, "test-data-" + cur_time_str + ".jsonl")
+    clean_data_and_save(test_data_set, output)
