--- examples/test_data_gen/test_data_gen_local/constants.py
@@ -1,2 +1,0 @@
-DOCUMENT_NODE = "document_node"
-TEXT_CHUNK = "text_chunk"
