--- examples/test_data_gen/test_data_gen_pipeline/conda.yml
@@ -1,9 +1,0 @@
-name: test_data_gen_conda_env
-channels:
-  - defaults
-dependencies:
-  - python=3.10.12
-  - pip=23.2.1
-  - pip:
-    - mldesigner==0.1.0b17
-    - llama_index
